backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
torch
torch_geometric
requests
joblib
python-dotenv